fix IntelliJ IDEA Inspection issues

+++ core/src/main/java/com/alibaba/fastjson2/JSONObject.java
@@ -1002,12 +1002,12 @@ public class JSONObject
             }
 
             if (value instanceof Float) {
-                float floatValue = ((Float) value).floatValue();
+                float floatValue = (Float) value;
                 return toBigDecimal(floatValue);
             }
 
             if (value instanceof Double) {
-                double doubleValue = ((Double) value).doubleValue();
+                double doubleValue = (Double) value;
                 return toBigDecimal(doubleValue);
             }
