+++ core/src/main/java/com/alibaba/fastjson2/writer/FieldWriterFloatField.java
@@ -55,7 +55,7 @@ class FieldWriterFloatField<T>
 
         writeFieldName(jsonWriter);
 
-        float floatValue = value.floatValue();
+        float floatValue = value;
         if (decimalFormat != null) {
             jsonWriter.writeFloat(floatValue, decimalFormat);
         } else {
@@ -72,7 +72,7 @@ class FieldWriterFloatField<T>
         if (value == null) {
             jsonWriter.writeNumberNull();
         } else {
-            float floatValue = value.floatValue();
+            float floatValue = value;
             if (decimalFormat != null) {
                 jsonWriter.writeFloat(floatValue, decimalFormat);
             } else {
